+++ deepinv/tests/test_sampling.py
@@ -130,8 +130,7 @@ def test_algo(name_algo, device):
     elif name_algo == "DPS":
         f = DPS(
             dinv.models.DiffUNet().to(device),
-            likelihood,
-            max_iter=5,
+            num_steps=5,
             verbose=False,
             device=device,
         )
@@ -145,9 +144,7 @@ def test_algo(name_algo, device):
 
 @pytest.mark.parametrize("name_algo", ["DiffPIR", "DPS", "DDRM"])
 def test_algo_inpaint(name_algo, device):
-    from deepinv.models import DiffUNet
-
-    x = torch.ones((1, 3, 32, 32)).to(device) / 2.0
+    x = torch.ones((1, 3, 32, 32)).to(device)
     x[:, 0, ...] = 0  # create a colored image
 
     torch.manual_seed(10)
@@ -159,7 +156,7 @@ def test_algo_inpaint(name_algo, device):
 
     y = physics(x)
 
-    model = DiffUNet().to(device)
+    model = dinv.models.DRUNet(device=device)
     likelihood = L2()
 
     if name_algo == "DiffPIR":
@@ -167,7 +164,9 @@ def test_algo_inpaint(name_algo, device):
             model, likelihood, max_iter=20, verbose=False, device=device, sigma=0.01
         )
     elif name_algo == "DPS":
-        algorithm = DPS(model, likelihood, max_iter=100, verbose=False, device=device)
+        algorithm = DPS(
+            model, num_steps=50, weight=2.0, alpha=0.01, verbose=False, device=device
+        )
     elif name_algo == "DDRM":
         algorithm = DDRM(model)
 
@@ -184,7 +183,7 @@ def test_algo_inpaint(name_algo, device):
 
     masked_target = x[mask]
     mean_target_masked = masked_target.mean()
-    mean_target_inmask = 1 / 3.0
+    mean_target_inmask = 2 / 3.0
 
     assert (mean_target_inmask - mean_crop).abs() < 0.2
     assert (mean_target_masked - mean_outside_crop).abs() < 0.02